Introduce task class + small refactoring
